Add files via upload
We will send Invalid JSON Data to see the Azure function working fine or not. This invalid Data should be sent to rejected folder once it is triggered with azure function. Once we update or any data in landing storage the azure function will check the data, Code will run to check if the data is valid or not.